v2: Brief Outline of Design  with v0.dev
In later versions, we will integrate the structure of the old personal website into this one. v0.dev was intended as a help guide for migrating the website from vanilla HTML+JS to a more advanced and nice-looking website with React. This is also my plan to quickly learn something about react, which seem to me that the general structures are similar.